requirements.txt
asgiref==3.5.2
countryinfo==0.1.2
Django==4.0.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-user-agents==0.4.0
djangorestframework==3.13.1
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
ua-parser==0.15.0
user-agents==2.2.0
